Add group relation

+++ apps/rpc/src/modules/contest/contest-router.ts
@@ -2,6 +2,7 @@ import { ContestSchema, ContestWriteSchema, ContestantSchema } from "@dotkomonli
 import type { inferProcedureInput, inferProcedureOutput } from "@trpc/server"
 import { z } from "zod"
 import { isCommitteeMember } from "../../authorization"
+import { ForbiddenError } from "../../error"
 import { withAuditLogEntry, withAuthentication, withAuthorization, withDatabaseTransaction } from "../../middlewares"
 import { PaginateInputSchema } from "@dotkomonline/utils"
 import { procedure, t } from "../../trpc"
@@ -15,6 +16,12 @@ const createContestProcedure = procedure
   .use(withDatabaseTransaction())
   .use(withAuditLogEntry())
   .mutation(async ({ input, ctx }) => {
+    const allowedGroups = ctx.authorizationService.intersectGroupAffiliations(ctx.principal.affiliations, [
+      input.groupId,
+    ])
+    if (allowedGroups.size === 0) {
+      throw new ForbiddenError(`User(ID=${ctx.principal.subject}) is not a member of Group(ID=${input.groupId})`)
+    }
     return ctx.contestService.create(ctx.handle, input)
   })
 
@@ -32,6 +39,13 @@ const updateContestProcedure = procedure
   .use(withDatabaseTransaction())
   .use(withAuditLogEntry())
   .mutation(async ({ input, ctx }) => {
+    const contest = await ctx.contestService.getById(ctx.handle, input.id)
+    const allowedGroups = ctx.authorizationService.intersectGroupAffiliations(ctx.principal.affiliations, [
+      contest.groupId,
+    ])
+    if (allowedGroups.size === 0) {
+      throw new ForbiddenError(`User(ID=${ctx.principal.subject}) is not authorized to update Contest(ID=${input.id})`)
+    }
     return ctx.contestService.update(ctx.handle, input.id, input.input)
   })
 
@@ -48,6 +62,13 @@ const deleteContestProcedure = procedure
   .use(withDatabaseTransaction())
   .use(withAuditLogEntry())
   .mutation(async ({ input, ctx }) => {
+    const contest = await ctx.contestService.getById(ctx.handle, input.id)
+    const allowedGroups = ctx.authorizationService.intersectGroupAffiliations(ctx.principal.affiliations, [
+      contest.groupId,
+    ])
+    if (allowedGroups.size === 0) {
+      throw new ForbiddenError(`User(ID=${ctx.principal.subject}) is not authorized to delete Contest(ID=${input.id})`)
+    }
     return ctx.contestService.delete(ctx.handle, input.id)
   })
 
@@ -83,6 +104,15 @@ const setWinnerProcedure = procedure
   .use(withDatabaseTransaction())
   .use(withAuditLogEntry())
   .mutation(async ({ input, ctx }) => {
+    const contest = await ctx.contestService.getById(ctx.handle, input.contestId)
+    const allowedGroups = ctx.authorizationService.intersectGroupAffiliations(ctx.principal.affiliations, [
+      contest.groupId,
+    ])
+    if (allowedGroups.size === 0) {
+      throw new ForbiddenError(
+        `User(ID=${ctx.principal.subject}) is not authorized to set winner for Contest(ID=${input.contestId})`
+      )
+    }
     return ctx.contestService.setWinner(ctx.handle, input.contestId, input.contestantId)
   })
 
@@ -102,6 +132,15 @@ const addContestantProcedure = procedure
   .use(withDatabaseTransaction())
   .use(withAuditLogEntry())
   .mutation(async ({ input, ctx }) => {
+    const contest = await ctx.contestService.getById(ctx.handle, input.contestId)
+    const allowedGroups = ctx.authorizationService.intersectGroupAffiliations(ctx.principal.affiliations, [
+      contest.groupId,
+    ])
+    if (allowedGroups.size === 0) {
+      throw new ForbiddenError(
+        `User(ID=${ctx.principal.subject}) is not authorized to add contestants to Contest(ID=${input.contestId})`
+      )
+    }
     return ctx.contestService.addContestant(
       ctx.handle,
       { contestId: input.contestId, userId: input.userId, resultValue: null },
@@ -123,6 +162,16 @@ const removeContestantProcedure = procedure
   .use(withDatabaseTransaction())
   .use(withAuditLogEntry())
   .mutation(async ({ input, ctx }) => {
+    const contestant = await ctx.handle.contestant.findUniqueOrThrow({ where: { id: input.id } })
+    const contest = await ctx.contestService.getById(ctx.handle, contestant.contestId)
+    const allowedGroups = ctx.authorizationService.intersectGroupAffiliations(ctx.principal.affiliations, [
+      contest.groupId,
+    ])
+    if (allowedGroups.size === 0) {
+      throw new ForbiddenError(
+        `User(ID=${ctx.principal.subject}) is not authorized to remove contestants from Contest(ID=${contest.id})`
+      )
+    }
     return ctx.contestService.removeContestant(ctx.handle, input.id)
   })
 
@@ -140,6 +189,16 @@ const updateContestantResultProcedure = procedure
   .use(withDatabaseTransaction())
   .use(withAuditLogEntry())
   .mutation(async ({ input, ctx }) => {
+    const contestant = await ctx.handle.contestant.findUniqueOrThrow({ where: { id: input.id } })
+    const contest = await ctx.contestService.getById(ctx.handle, contestant.contestId)
+    const allowedGroups = ctx.authorizationService.intersectGroupAffiliations(ctx.principal.affiliations, [
+      contest.groupId,
+    ])
+    if (allowedGroups.size === 0) {
+      throw new ForbiddenError(
+        `User(ID=${ctx.principal.subject}) is not authorized to update results for Contest(ID=${contest.id})`
+      )
+    }
     return ctx.contestService.updateContestantResult(ctx.handle, input.id, input.resultValue)
   })
